Add live updating of node edges during move
This commit adds the feature of "live-updating" the node edges when a
node moves. During the move event, the connections between the nearest
handles on the from and to nodes are updated dynamically, so the graph
edge appers to follow the node as it moves.

The key change was to make the displacement of the node as a result of
the drag-move be a property of the node data, rather than be a property
of the Vue component. This allows the current position of the node
during drag-move to be considered during the calculation of the nearest
pair of handles.